Add Transparency Groups to the C and CPP APIs
This also tweaks the DrawContext returned by new transparency group to be
a pointer to make it consistant with the other similar functions. Tests
updated to account for the change.

The transparency group is modified so it can be constructed with it's
entire bounding box and not just a width and a height. This is because
the paint is transformed by the current transform coordinate space.

Added python test for new functionality.

Fixes https://github.com/jpakkane/capypdf/issues/31